instrumentation/python/aws-lambda/hello_world/requirements.txt
requests
# add this package to ensure trace context is added to logs
opentelemetry.instrumentation.logging==0.54b1
